+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Updated
 - Scoped MonoBehaviours
 - MonoBehaviour prefabs registration by prefab name from Resources
+- ScriptableObjectConfiguration as configuration source
 
 ## [2.0.0] - 2022-07-21
 ### Updated